Codeforces Round 1080 (Div. 3)

+++ competition/Codeforces/Educational Codeforces Round 186 (Rated for Div. 2)/hello.cpp
@@ -1,14 +1,14 @@
 #include <bits/stdc++.h>
 using namespace std;
 #define int long long
-typedef pair<int, int> PII ;
+typedef pair<int, int> PII;
 typedef pair<int, PII> PIII;
 #define pi acos(-1)
-using ll = long long;
+using int = long long;
 const int mod = 1e10 + 7;
 const long long INF = LLONG_MAX;
 #define all(v) (v).begin, (v).end
 
-int main() {
-
+int main()
+{
 }
